requirements.txt
cliff==2.3.0
datakit-core==0.2.1
python-gitlab==2.5.0
